domain to schema
