--- packages/models-library/src/models_library/projects_nodes_ui.py
@@ -1,21 +1,0 @@
-"""
-    Models node UI (legacy model, use instead projects.ui.py)
-"""
-
-from typing import Annotated
-
-from pydantic import BaseModel, ConfigDict, Field, PlainSerializer
-from pydantic_extra_types.color import Color
-
-
-class PositionUI(BaseModel):
-    x: int = Field(..., description="The x position", examples=[["12"]])
-    y: int = Field(..., description="The y position", examples=[["15"]])
-
-    model_config = ConfigDict(extra="forbid")
-
-
-class MarkerUI(BaseModel):
-    color: Annotated[Color, PlainSerializer(Color.as_hex), Field(...)]
-
-    model_config = ConfigDict(extra="forbid")
