+++ services/api-server/src/simcore_service_api_server/models/domain/projects.py
@@ -3,14 +3,12 @@
 from models_library.projects_access import AccessRights
 from models_library.projects_nodes import InputTypes, Node, OutputTypes
 from models_library.projects_nodes_io import SimCoreFileLink
-from models_library.projects_ui import StudyUI
 
 assert AccessRights  # nosec
 assert InputTypes  # nosec
 assert Node  # nosec
 assert OutputTypes  # nosec
 assert SimCoreFileLink  # nosec
-assert StudyUI  # nosec
 
 __all__: tuple[str, ...] = (
     "AccessRights",
@@ -18,5 +16,4 @@ __all__: tuple[str, ...] = (
     "Node",
     "OutputTypes",
     "SimCoreFileLink",
-    "StudyUI",
 )
